finpilot-backend-main/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic[email]==2.9.2
firebase-admin==6.5.0
bcrypt==4.2.0
